requirements.txt
blinker==1.9.0
boto3==1.26.87
botocore==1.29.165
click==8.1.8
colorama==0.4.6
Flask==2.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
playsound==1.2.2
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
s3transfer==0.6.2
simpleaudio==1.0.4
six==1.17.0
urllib3==1.26.20
websockets==15.0.1
Werkzeug==3.1.3